tools/requirements.txt
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
twilio>=9.0.0
